Refactor settings pannel, add flare to contact tracing confirmation screen
