+++ src/Patient/Settings/PersonalInformation.js
@@ -37,7 +37,7 @@ const useStyles = makeStyles({
     },
 })
 
-const LanguageQuestion = () => {
+const PersonalInformation = () => {
     const classes = useStyles();
     const { patientUIStore } = useStores();
     const { t, i18n } = useTranslation('translation');
@@ -49,12 +49,9 @@ const LanguageQuestion = () => {
                 <Typography variant="h2">{t("patient.profile.personalInfo")}</Typography>
             </div>
             <Button onClick={patientUIStore.goToPasswordUpdate} >{t("patient.profile.changePassword")}</Button>
-            {/*
-            <Button onClick={() => { }} >{t("patient.profile.editNotifications") } (Coming Soon)</Button>
-            <Button onClick={() => { }} >{t("patient.profile.changeUsername")} (Coming Soon)</Button>
-            */}
+            <Button onClick={patientUIStore.goToContactTracingUpdate} >{t('updatedContactTracing.button')}</Button>
         </div>
     );
 }
 
-export default LanguageQuestion;
+export default PersonalInformation;
